Rename new DA dir and files based on review
