+++ .vscode/launch.json
@@ -35,13 +35,13 @@
 			"preLaunchTask": "npm"
 		},
 		{
-			"name": "Launch as server DlvDap",
+			"name": "Launch as server (dlv dap)",
 			"type": "node",
 			"protocol": "inspector",
 			"request": "launch",
-			"program": "${workspaceFolder}/out/src/debugAdapterDlvDap/goDlvDapDebugMain.js",
+			"program": "${workspaceFolder}/out/src/debugAdapter2/goDlvDebugMain.js",
 			"args": [
-				"--server=4722"
+				"--server=4711"
 			],
 			"sourceMaps": true,
 			"outFiles": [
