+++ package.json
@@ -64,7 +64,7 @@
     "@types/fs-extra": "^8.1.0",
     "@types/glob": "^7.1.1",
     "@types/mocha": "^7.0.2",
-    "@types/node": "^13.13.11",
+    "@types/node": "^13.13.1",
     "@types/semver": "^7.1.0",
     "@types/sinon": "^9.0.0",
     "@types/vscode": "^1.41.0",
@@ -755,7 +755,7 @@
       {
         "type": "godlvdap",
         "label": "Go Dlv Dap (Experimental)",
-        "program": "./out/src/debugAdapterDlvDap/goDlvDapDebugMain.js",
+        "program": "./out/src/debugAdapter2/goDlvDebugMain.js",
         "runtime": "node",
         "languages": [
           "go"
